requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1
rapidfuzz==3.9.7
requests
reportlab==4.0.8
PyJWT==2.8.0
apscheduler==3.10.4
bcrypt==4.1.2